# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  description: The microservice CI jobs API.
  version: 0.0.1

components:
    schemas:
        Command:
            properties:
                program:
                    type: string
                    description: The program to launch for this command
                arguments:
                    type: array
                    description: The list of arguments for this command
                    items:
                        type: string
        Job:
            properties:
                project:
                    type: string
                    description: The ID of the project for this job.
                script:
                    type: array
                    description: The list of commands to execute for this job.
                    items:
                        $ref: '#/components/schemas/Command'

paths:
    /jobs:
        # SEARCH FOR JOBS
        get:
            description: Search for jobs.
            responses:
                '200':
                    description: Request successfull
                    content: 
                        application/json:
                            schema:
                                type: array
                                description: An array of jobs.
                                items:
                                    type: object
                                    $ref: '#/components/schemas/Job'
                '500':
                    description: Something wrong happened.
        
        # CREATE NEW JOB.
        post:
            description: Create a new job.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Job'
            responses:
                '201':
                    description: Job was created.
                    headers:
                        Location:
                            type: string
                            description: The link to the created job.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Job'
                '403':
                    description: Job already exists for this project.
                '5XX':
                    description: Something wrong happened

    /jobs/{id}:
        get:
            description: Get a job.
            responses:
                '200':
                    description: Job was found.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Job'
                '404':
                    description: Job does not exist
                '5XX':
                    description: Something wrong happened
        patch:
            description: Update a job
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Job'
            responses:
                '404':
                    description: Job does not exist.
                '5XX':
                    description: Something wrong happened
        delete:
            description: Delete a job
            responses:
                '204':
                    description: Job was deleted
                '404':
                    description: Job does not exist
                '5XX':
                    description: Something wrong happened

    /jobs/{id}/command:
        post:
            description: Add a command for a job.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Command'
            responses:
                '201':
                    description: Command was added to job
                '404':
                    description: Job does not exist
                '5XX':
                    description: Something wrong happened
    /rpc/{projectId}/execute:
        post:
            description: Build hook to send commands to docker-runner microservice
            requestBody:
                required: true
                content:
                    application/json:
                        description: Any data that you want to send to docker-runner along with commands
                response:
                    '200':
                        description: The payload has been processed
                    '5XX':
                        description: Something wrong happened